refactor(ui): share 3D chart option shell
